♻️ Extract public/duration.go from duration.go and add aliases
